+++ RELEASENOTES.md
@@ -92,6 +92,9 @@
     *   Parse "f800" as channel count of 5 for Dolby in DASH manifest
         ([#688](https://github.com/androidx/media/issues/688)).
 *   Smooth Streaming Extension:
+    *   Add experimental support for parsing subtitles during extraction. You
+        can enable this using
+        `SsMediaSource.Factory.experimentalParseSubtitlesDuringExtraction()`.
 *   RTSP Extension:
 *   Decoder Extensions (FFmpeg, VP9, AV1, MIDI, etc.):
     *   MIDI decoder: Ignore SysEx event messages